Замена SightList на SliverSightList
